+++ c3smembership/data/repository/general_assembly_repository.py
@@ -3,6 +3,8 @@
 Repository for accessing and operating with member data.
 """
 
+from datetime import date
+
 from sqlalchemy import (
     and_,
     or_,
@@ -10,6 +12,35 @@ from sqlalchemy import (
 
 from c3smembership.data.model.base import DBSession
 from c3smembership.data.model.base.c3smember import C3sMember
+
+
+GENERAL_ASSEMBLIES = [
+    {
+        'number': '1',
+        'name': '1. ordentliche Generalversammlung',
+        'date': date(2014, 8, 23),
+    },
+    {
+        'number': '2',
+        'name': '2. ordentliche Generalversammlung',
+        'date': date(2015, 6, 13),
+    },
+    {
+        'number': '3',
+        'name': '3. ordentliche Generalversammlung',
+        'date': date(2016, 4, 17),
+    },
+    {
+        'number': '4',
+        'name': '4. ordentliche Generalversammlung',
+        'date': date(2017, 4, 2),
+    },
+    {
+        'number': '5',
+        'name': '5. ordentliche Generalversammlung',
+        'date': date(2018, 6, 3),
+    },
+]
 
 
 class GeneralAssemblyRepository(object):
@@ -33,6 +64,7 @@ class GeneralAssemblyRepository(object):
         Returns:
           a list of *n* member objects
         """
+        # pylint: disable=no-member
         return DBSession.query(C3sMember).filter(
             and_(
                 C3sMember.is_member_filter(),
@@ -56,5 +88,51 @@ class GeneralAssemblyRepository(object):
         Returns:
             object: C3sMember object
         """
+        # pylint: disable=no-member
         return DBSession.query(C3sMember).filter(
             C3sMember.email_invite_token_bcgv18 == token).first()
+
+    @classmethod
+    def get_member_invitations(cls, membership_number):
+        """
+        Get all general assembly invitations of the member.
+
+        Args:
+            membership_number: The membership number of the member of which
+                the general assembly invitations are returned.
+
+        Returns:
+            All general assemblies with number, name, date, invited flag and
+            sent date to which the member could have been invited according to
+            their membership status.
+        """
+        result = []
+        # pylint: disable=no-member
+        member = DBSession.query(C3sMember) \
+            .filter(C3sMember.membership_number == membership_number) \
+            .first()
+        flags = [
+            member.email_invite_flag_bcgv14,
+            member.email_invite_flag_bcgv15,
+            member.email_invite_flag_bcgv16,
+            member.email_invite_flag_bcgv17,
+            member.email_invite_flag_bcgv18,
+        ]
+        sent = [
+            member.email_invite_date_bcgv14,
+            member.email_invite_date_bcgv15,
+            member.email_invite_date_bcgv16,
+            member.email_invite_date_bcgv17,
+            member.email_invite_date_bcgv18,
+        ]
+        # pylint: disable=consider-using-enumerate
+        for i in range(len(GENERAL_ASSEMBLIES)):
+            if member.is_member(GENERAL_ASSEMBLIES[i]['date']):
+                result.append({
+                    'number': GENERAL_ASSEMBLIES[i]['number'],
+                    'name': GENERAL_ASSEMBLIES[i]['name'],
+                    'date': GENERAL_ASSEMBLIES[i]['date'],
+                    'flag': flags[i],
+                    'sent': sent[i],
+                    })
+        return result