+++ c3smembership/presentation/views/membership_member_detail.py
@@ -17,6 +17,8 @@ from c3smembership.data.model.base.dues15invoice import Dues15Invoice
 from c3smembership.data.model.base.dues16invoice import Dues16Invoice
 from c3smembership.data.model.base.dues17invoice import Dues17Invoice
 from c3smembership.data.model.base.dues18invoice import Dues18Invoice
+from c3smembership.data.repository.general_assembly_repository import \
+    GeneralAssemblyRepository
 
 LOG = logging.getLogger(__name__)
 
@@ -25,12 +27,17 @@ def get_member_details(request, member):
     """
     Gets the member details.
     """
+    shares = request.registry.share_information.get_member_shares(
+        member.membership_number)
     invoices15 = Dues15Invoice.get_by_membership_no(member.membership_number)
     invoices16 = Dues16Invoice.get_by_membership_no(member.membership_number)
     invoices17 = Dues17Invoice.get_by_membership_no(member.membership_number)
     invoices18 = Dues18Invoice.get_by_membership_no(member.membership_number)
-    shares = request.registry.share_information.get_member_shares(
-        member.membership_number)
+    general_assembly_invitations = sorted(
+        GeneralAssemblyRepository.get_member_invitations(
+            member.membership_number),
+        key=lambda ga: ga['date'],
+        reverse=True)
 
     return {
         'today': date.today().strftime('%Y-%m-%d'),
@@ -41,6 +48,7 @@ def get_member_details(request, member):
         'invoices16': invoices16,
         'invoices17': invoices17,
         'invoices18': invoices18,
+        'general_assembly_invitations': general_assembly_invitations,
     }